+++ utils/wallet/provider/uphold/uphold.go
@@ -224,7 +224,7 @@ type createCardRequest struct {
 }
 
 // IsUserKYC - is this user a "member"
-func (w *Wallet) IsUserKYC(ctx context.Context) bool {
+func (w *Wallet) IsUserKYC(ctx context.Context) (bool, error) {
 	// get logger
 	logger, err := appctx.GetLogger(ctx)
 	if err != nil {
@@ -238,12 +238,12 @@ func (w *Wallet) IsUserKYC(ctx context.Context) bool {
 	gwPublicKey, err := hex.DecodeString(grantWalletPublicKey)
 	if err != nil {
 		logger.Error().Err(err).Msg("invalid system public key")
-		return false
+		return false, fmt.Errorf("invalid system public key: %w", err)
 	}
 	gwPrivateKey, err := hex.DecodeString(grantWalletPrivateKey)
 	if err != nil {
 		logger.Error().Err(err).Msg("invalid system private key")
-		return false
+		return false, fmt.Errorf("invalid system private key: %w", err)
 	}
 
 	grantWallet := Wallet{
@@ -256,20 +256,21 @@ func (w *Wallet) IsUserKYC(ctx context.Context) bool {
 		PubKey:  httpsignature.Ed25519PubKey([]byte(gwPublicKey)),
 	}
 
-	// create a transaction
+	// prepare a transaction by creating a payload
 	transactionB64, err := grantWallet.PrepareTransaction(altcurrency.BAT, decimal.New(0, 1), w.Info.ProviderID, "")
 	if err != nil {
 		logger.Error().Err(err).Msg("failed to prepare transaction")
-		return false
+		return false, fmt.Errorf("failed to prepare transaction: %w", err)
 	}
 
+	// submit the transaction the payload
 	uhResp, err := grantWallet.SubmitTransaction(transactionB64, false)
 	if err != nil {
 		logger.Error().Err(err).Msg("failed to submit transaction")
-		return false
+		return false, fmt.Errorf("failed to submit transaction: %w", err)
 	}
 
-	return uhResp.KYC
+	return uhResp.KYC, nil
 }
 
 // sign registration for this wallet with Uphold with label
