+++ wallet/datastore.go
@@ -1,27 +1,36 @@
 package wallet
 
 import (
-	"context"
 	"errors"
-	"net/http"
 	"os"
-	"time"
 
 	"github.com/brave-intl/bat-go/datastore/grantserver"
 	"github.com/brave-intl/bat-go/utils/altcurrency"
 	errorutils "github.com/brave-intl/bat-go/utils/errors"
-	"github.com/brave-intl/bat-go/utils/handlers"
 	"github.com/brave-intl/bat-go/utils/wallet"
 	walletutils "github.com/brave-intl/bat-go/utils/wallet"
 	"github.com/getsentry/sentry-go"
 	"github.com/jmoiron/sqlx"
+	"github.com/prometheus/client_golang/prometheus"
 	"github.com/rs/zerolog/log"
 	uuid "github.com/satori/go.uuid"
-	"github.com/shopspring/decimal"
 
 	// needed for magic migration
 	_ "github.com/golang-migrate/migrate/v4/source/file"
 )
+
+var (
+	tooManyCardsCounter = prometheus.NewCounter(
+		prometheus.CounterOpts{
+			Name:        "too_many_linked_cards",
+			Help:        "A counter for too many linked cards",
+			ConstLabels: prometheus.Labels{"service": "wallet"},
+		})
+)
+
+func init() {
+	prometheus.MustRegister(tooManyCardsCounter)
+}
 
 // Datastore holds the interface for the wallet datastore
 type Datastore interface {
@@ -85,7 +94,6 @@ func NewPostgres() (Datastore, ReadOnlyDatastore, error) {
 	walletDB, err := NewWritablePostgres("", true)
 	if err != nil {
 		sentry.CaptureException(err)
-		sentry.Flush(time.Second * 2)
 		log.Panic().Err(err).Msg("Must be able to init postgres connection to start")
 		return nil, nil, err
 	}
@@ -94,7 +102,6 @@ func NewPostgres() (Datastore, ReadOnlyDatastore, error) {
 		walletRODB, err = NewReadOnlyPostgres(roDB, false)
 		if err != nil {
 			sentry.CaptureException(err)
-			sentry.Flush(time.Second * 2)
 			log.Error().Err(err).Msg("Could not start reader postgres connection")
 			return nil, nil, err
 		}
@@ -129,7 +136,7 @@ func (pg *Postgres) UpsertWallet(wallet *wallet.Info) error {
 
 // GetWallet by ID
 func (pg *Postgres) GetWallet(ID uuid.UUID) (*wallet.Info, error) {
-	statement := "select * from wallets where id = $1"
+	statement := "select id, provider, provider_id, public_key, provider_linking_id, anonymous_address from wallets where id = $1"
 	wallets := []wallet.Info{}
 	err := pg.RawDB().Select(&wallets, statement, ID)
 	if err != nil {
@@ -151,7 +158,7 @@ func (pg *Postgres) GetWallet(ID uuid.UUID) (*wallet.Info, error) {
 // GetWalletByPublicKey gets a wallet by a public key
 func (pg *Postgres) GetWalletByPublicKey(pk string) (*walletutils.Info, error) {
 	statement := `
-	SELECT *
+	SELECT id, provider, provider_id, public_key, provider_linking_id, anonymous_address
 	FROM wallets
 	WHERE public_key = $1
 	`
@@ -163,7 +170,7 @@ func (pg *Postgres) GetWalletByPublicKey(pk string) (*walletutils.Info, error) {
 // GetByProviderLinkingID gets a wallet by a provider address
 func (pg *Postgres) GetByProviderLinkingID(providerLinkingID uuid.UUID) (*[]walletutils.Info, error) {
 	statement := `
-	SELECT *
+	SELECT id, provider, provider_id, public_key, provider_linking_id, anonymous_address
 	FROM wallets
 	WHERE provider_linking_id = $1
 	`
@@ -209,60 +216,6 @@ func (pg *Postgres) SetAnonymousAddress(ID string, anonymousAddress *uuid.UUID) 
 	return err
 }
 
-// LinkWallet links a wallet and transfers funds to newly linked wallet
-func (service *Service) LinkWallet(
-	ctx context.Context,
-	info *walletutils.Info,
-	transaction string,
-	anonymousAddress *uuid.UUID,
-) error {
-	// do not confirm this transaction yet
-	tx, err := service.SubmitCommitableAnonCardTransaction(
-		ctx,
-		info,
-		transaction,
-		"",
-		false,
-	)
-	if err != nil {
-		return handlers.WrapError(err, "unable to verify transaction", http.StatusBadRequest)
-	}
-	if tx.UserID == "" {
-		err := errors.New("user id not provided")
-		return handlers.WrapError(err, "unable to link wallet", http.StatusBadRequest)
-	}
-	providerLinkingID := uuid.NewV5(walletClaimNamespace, tx.UserID)
-	if info.ProviderLinkingID != nil {
-		// check if the member matches the associated member
-		if !uuid.Equal(*info.ProviderLinkingID, providerLinkingID) {
-			return handlers.WrapError(errors.New("wallets do not match"), "unable to match wallets", http.StatusForbidden)
-		}
-		if anonymousAddress != nil && info.AnonymousAddress != nil && !uuid.Equal(*anonymousAddress, *info.AnonymousAddress) {
-			err := service.Datastore.SetAnonymousAddress(info.ID, anonymousAddress)
-			if err != nil {
-				return handlers.WrapError(err, "unable to set anonymous address", http.StatusInternalServerError)
-			}
-		}
-	} else {
-		err := service.Datastore.LinkWallet(info.ID, providerLinkingID, anonymousAddress)
-		if err != nil {
-			status := http.StatusInternalServerError
-			if err == ErrTooManyCardsLinked {
-				status = http.StatusConflict
-			}
-			return handlers.WrapError(err, "unable to link wallets", status)
-		}
-	}
-
-	if decimal.NewFromFloat(0).LessThan(tx.Probi) {
-		_, err := service.SubmitCommitableAnonCardTransaction(ctx, info, transaction, "", true)
-		if err != nil {
-			return handlers.WrapError(err, "unable to transfer tokens", http.StatusBadRequest)
-		}
-	}
-	return nil
-}
-
 // TxLinkWalletInfo pass a tx to set the anonymous address
 func (pg *Postgres) TxLinkWalletInfo(tx *sqlx.Tx, ID string, providerLinkingID uuid.UUID, anonymousAddress *uuid.UUID) error {
 	statement := `
@@ -271,18 +224,22 @@ func (pg *Postgres) TxLinkWalletInfo(tx *sqlx.Tx, ID string, providerLinkingID u
 			provider_linking_id = $2,
 			anonymous_address = $3
 	WHERE id = $1;`
-	_, err := tx.Exec(
+	r, err := tx.Exec(
 		statement,
 		ID,
 		providerLinkingID,
 		anonymousAddress,
 	)
+	count, _ := r.RowsAffected()
+	if count < 1 {
+		return errors.New("should have updated at least one wallet")
+	}
 	return err
 }
 
 func txGetByProviderLinkingID(tx *sqlx.Tx, providerLinkingID uuid.UUID) (*[]walletutils.Info, error) {
 	statement := `
-	SELECT *
+	SELECT id, provider, provider_id, public_key, provider_linking_id, anonymous_address
 	FROM wallets
 	WHERE provider_linking_id = $1
 	`
@@ -316,7 +273,7 @@ func (pg *Postgres) LinkWallet(ID string, providerLinkingID uuid.UUID, anonymous
 					"providerLinkingId": providerLinkingID.String(),
 					"anonymousAddress":  anonAddr,
 				})
-				sentry.CaptureMessage("too many cards linked")
+				tooManyCardsCounter.Inc()
 			})
 		}
 		return ErrTooManyCardsLinked
